pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bids-data-pipeline"
version = "0.1.0"
description = "Research pipeline for SEC merger-background extraction."
readme = "docs/design.md"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.49",
  "edgartools>=5.23",
  "openpyxl>=3.1",
  "pydantic>=2.0",
  "pytest>=8.0",
]

[project.scripts]
pipeline = "pipeline.cli:main"

[tool.setuptools.packages.find]
include = ["pipeline*"]
